The apiRequest function separated from jokeService.
